Challenge finished

+++ src/thunks/index.ts
@@ -1,12 +1,13 @@
 import { Action } from 'redux'
-import { fetchTournamentsStart, fetchTournamentsSuccess, tournamentsAPIError, deleteTournamentStart, deleteTournamentSuccess } from '../actions'
+import { fetchTournamentsStart, fetchTournamentsSuccess, tournamentsAPIError, editTournamentStart, editTournamentSuccess, createTournamentStart, createTournamentSuccess, deleteTournamentStart, deleteTournamentSuccess } from '../actions'
 import { ITournament } from '../actions/tournaments.types'
 import { API_TOURNAMENTS_URL } from '../constants/api'
 
 import { RootState } from '../store'
 import { ThunkAction } from 'redux-thunk'
 
-type ThunkResult<R> = ThunkAction<R, RootState, unknown, Action>;
+type ThunkResult<R> = ThunkAction<R, RootState, unknown, Action>
+
 export const fetchTournamentsThunk = (): ThunkResult<void> => {
   return dispatch => {
 		dispatch(fetchTournamentsStart())
@@ -15,7 +16,7 @@ export const fetchTournamentsThunk = (): ThunkResult<void> => {
 		  .then(tournamentsData => tournamentsData.json())
 		  .then(tournamentsData => {
         setTimeout(() => { // To give it a realistic loading feel
-          // dispatch(fetchTournamentsError('forced error'))
+          // dispatch(tournamentsAPIError('forced error'))
   	      dispatch(fetchTournamentsSuccess(tournamentsData))
         }, 1000)
 		  })
@@ -27,6 +28,9 @@ export const fetchTournamentsThunk = (): ThunkResult<void> => {
 
 export const deleteTournamentThunk = (tournament: ITournament): ThunkResult<void> => {
   return dispatch => {
+    const confirm = window.confirm('Do you really want to delete this tournament?')
+    if(confirm == false) return
+
     dispatch(deleteTournamentStart(tournament))
 
     fetch(API_TOURNAMENTS_URL + '/' + tournament.id, {
@@ -38,5 +42,57 @@ export const deleteTournamentThunk = (tournament: ITournament): ThunkResult<void
 		  .catch(error => {
 				dispatch(tournamentsAPIError(error))
 		  })
+  }
+}
+
+export const editTournamentThunk = (tournament: ITournament): ThunkResult<void> => {
+  return dispatch => {
+    const newTournamentName = prompt('New Tournament Name')
+    if(newTournamentName){
+      dispatch(editTournamentStart(tournament, newTournamentName))
+
+      fetch(API_TOURNAMENTS_URL + '/' + tournament.id, {
+        method: 'PATCH',
+        headers: {
+          "Content-type": "application/json; charset=UTF-8"
+        },
+        body: JSON.stringify({
+          name: newTournamentName
+        })
+      })
+        .then(updatedTournament => updatedTournament.json())
+  		  .then((updatedTournament: ITournament) => {
+          dispatch(editTournamentSuccess(updatedTournament))
+  		  })
+  		  .catch(error => {
+  				dispatch(tournamentsAPIError(error))
+  		  })
+    }
+  }
+}
+
+export const createTournamentThunk = (): ThunkResult<void> => {
+  return dispatch => {
+    const tournamentName = prompt('Tournament Name')
+    if(tournamentName){
+      dispatch(createTournamentStart(tournamentName))
+
+      fetch(API_TOURNAMENTS_URL, {
+        method: 'POST',
+        headers: {
+          "Content-type": "application/json; charset=UTF-8"
+        },
+        body: JSON.stringify({
+          name: tournamentName
+        })
+      })
+        .then(newTournament => newTournament.json())
+  		  .then((newTournament: ITournament) => {
+          dispatch(createTournamentSuccess(newTournament))
+  		  })
+  		  .catch(error => {
+  				dispatch(tournamentsAPIError(error))
+  		  })
+    }
   }
 }
